feat: add house model

+++ modules/vehicle/vehicle.service.js
@@ -1,5 +1,6 @@
 import { uuidv7 } from "uuidv7";0
 import { VehicleModel } from "../../models/vehicleModel.js";
+import handleResponse from "../../utils/index.js";
 
 export const createNewVehicle = async (req, res) => {
   try {
@@ -45,18 +46,27 @@ export const createNewVehicle = async (req, res) => {
 
 export const getAllVehicle = async (req, res) => {
   try {
-    const vehicles = await VehicleModel.find(
+    const allVehicle = await VehicleModel.find(
       {},
-      { _id: 0, __v: 0, createdAt: 0, updatedAt: 0 }
+      { _id: 0, __v:   0, updatedAt: 0 }
     );
 
-    return res.status(200).json({
-      message: "Vehicles fetched successfully",
-      vehicles,
+    
+    if (!allVehicle) {
+    return handleResponse(res, 200, "No vehicle found", {
+      success: true,
+      data: {},
+    });
+    }
+    return handleResponse(res, 200, "Vehicle data fetched successfully", {
+      success: true,
+      data: allVehicle
     });
   } catch (error) {
-    console.error("Error fetching vehicles:", error);
-    return res.status(500).json({ message: "Internal server error" });
+    console.log( error);
+    return handleResponse(res, 500, "Internal server error", {
+      success: false,
+    });
   }
 };
 
@@ -84,12 +94,19 @@ export const getSingleVehicle = async (req, res) => {
 
 export const updateVehicle = async (req, res) => {
   try {
+   const  { id } = req.params;
     const { plateNumber, color, model, brand, type } = req.body;
 
+    if (!id) {
+      return handleResponse(res, 400, "Missing required field: Id", {
+        success: false,
+      });
+    }
+
     const updatedVehicle = await VehicleModel.findOneAndUpdate(
-      { plateNumber },
-      { color, model, brand, type },
-      { new: true, projection: { _id: 0, __v: 0, createdAt: 0, updatedAt: 0 } }
+      { id },
+      { color, model, brand, type, updated_at: new Date() },
+      { new: true, projection: { _id: 0, __v: 0, createdAt: 0 } }
     );
 
     if (!updatedVehicle) {
@@ -102,36 +119,43 @@ export const updateVehicle = async (req, res) => {
       message: "Vehicle updated successfully",
     });
   } catch (error) {
-    console.error("Error updating vehicle:", error);
-    return res.status(500).json({
-      message: "Failed to update vehicle",
-      error: error.message,
-    });
+    console.log(error);
+    return res.status(500).json({ message: error});
   }
 };
 
 export const deleteVehicle = async (req, res) => {
   try {
-    const { plateNumber } = req.body;
+    const { id } = req.params;
 
-    const deletedVehicle = await VehicleModel.findOneAndDelete(
-      { plateNumber }
-    );
-
-    if (!deletedVehicle) {
-      return res.status(404).json({
-        message: "Vehicle not found",
-      });
+    if (!id ) {
+      return handleResponse(
+        res,
+        400,
+        "Missing required field: platenumber",
+        {
+          success: false,
+        }
+      );
     }
 
-    return res.status(200).json({
-      message: "Vehicle deleted successfully",
-    });
+    const vehicleToDelete = await VehicleModel.findOneAndDelete(
+      { id: id },
+      {
+        options: { new: true},
+      }
+    );
+
+    if (!vehicleToDelete) {
+      return handleResponse(res, 400, "Vehicle not found", {success: false });
+    }
+return handleResponse(res, 200, "Vehicle deleted sucessfully", {
+success: true,
+});
   } catch (error) {
     console.error("Error deleting vehicle:", error);
-    return res.status(500).json({
-      message: "Failed to delete vehicle",
-      error: error.message,
-    });
+    return handleResponse(res, 200, "Internal server error", {
+success: false,
+    }); 
   }
 };
